+++ src/emotional_memory/__init__.py
@@ -101,6 +101,14 @@ with contextlib.suppress(ImportError):
 
     _sentence_transformers_available = True
 
+with contextlib.suppress(ImportError):
+    from emotional_memory.integrations.mem0 import (
+        EmotionalMemoryMem0Backend as EmotionalMemoryMem0Backend,
+    )
+    from emotional_memory.integrations.mem0 import (
+        messages_to_content as messages_to_content,
+    )
+
 _langchain_available = False
 with contextlib.suppress(ImportError):
     from emotional_memory.integrations.langchain import (
@@ -135,6 +143,7 @@ __all__ = [
     "EmotionalMemory",
     "EmotionalMemoryChatHistory",
     "EmotionalMemoryConfig",
+    "EmotionalMemoryMem0Backend",
     "EmotionalTag",
     "GenericAppraisalVector",
     "HeuristicQueryClassifier",
@@ -176,6 +185,7 @@ __all__ = [
     "hebbian_strengthen",
     "label_tag",
     "make_emotional_tag",
+    "messages_to_content",
     "recommended_conversation_policy",
     "spreading_activation",
     "store_all_messages",